Update turbosnap.md
I'd like to make a tweak to downplay TurboSnap a little by warning people about using it immediately. Hopefully this will safeguard against criticism about not enabling it out of the box as well as warning people of some of the pitfalls. Made a few other small tweaks along the way.